crush: fix distortion of straw scalers by 0-weight items
The presence of a 0-weight item in a straw bucket should have no effect
on the placement of other items.  Add a test validating that and fix
crush_calc_straw() to fix the distortion.

Note that this effects the *precalculation* of the straw bucket inputs and
does not effect the actually mapping process given a compiled or encoded
CRUSH map.  Backporting to old code will make the next bucket change
generate (more) correct straw scalers and likely trigger some mapping
changes, but will not lead to inconsistent mappings.

Backport: giant, firefly
Signed-off-by: Sage Weil <sage@redhat.com>